pyproject.toml
[project]
name = "mysql-character-update"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql==0.3.2",
    "python-dotenv>=1.0.0",
]
